requirements.txt
django
django-cors-headers
python-dotenv
djangorestframework
djangorestframework-simplejwt
